Ono2_NewsBot/requirements.txt
beautifulsoup4==4.4.1
pymorphy2==0.8
python-telegram-bot==5.2.0
nltk==3.2.2
